backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.11.1
psycopg2-binary>=2.9.6

# Authentication and security
python-jose>=3.3.0
passlib>=1.7.4
siwe>=2.0.0

# Web3 and blockchain
web3>=6.5.0
eth-account>=0.8.0

# HTTP client
aiohttp>=3.8.5

# Redis for rate limiting
redis>=4.6.0

# Utilities
python-dotenv>=1.0.0